Development, added readme, removed configs

+++ main.py
@@ -2,6 +2,7 @@ import configparser
 import datetime
 import logging
 import os
+import platform
 import sys
 import time
 import solar_edge
@@ -14,8 +15,8 @@ SOLAR_EDGE_API_END_TIME = datetime.datetime.now()
 DEVICES_CONFIGURATION = []
 DEVICES = []
 
-CONFIG_FILE_NAME = "ustawienia.cfg"
-DEVICE_DATA_FILE_NAME = "urzadzenia.csv"
+CONFIG_FILE_NAME = "settings.cfg"
+DEVICE_DATA_FILE_NAME = "devices.csv"
 LOGNAME = "smartbedkow.log"
 
 
@@ -23,12 +24,14 @@ def get_app_configuration_from_file():
     logging.info("Reading configuration from %s..." % CONFIG_FILE_NAME)
     config = configparser.ConfigParser()
     config.read(CONFIG_FILE_NAME)
-    solar_edge_api_key = config["USTAWIENIA"]["solar_edge_api_key"]
-    wattage_threshold = config["USTAWIENIA"]["minimalna_produkcja_do_wlaczenia_grzejnikow"]
+    solar_edge_api_key = config["SETTINGS"]["solar_edge_api_key"]
+    solar_edge_site_id = config["SETTINGS"]["solar_edge_site_id"]
+    wattage_threshold = config["SETTINGS"]["minimal_power_produced_to_turn_on_radiators_in_watts"]
     logging.info("Wattage threshold read from configuration file: %s W" % wattage_threshold)
-    refresh_interval = config["USTAWIENIA"]["czestotliwosc_odswiezania_w_sekundach"]
+    refresh_interval = config["SETTINGS"]["refresh_rate_in_seconds"]
     logging.info("Refresh interval read from configuration file: %s s" % refresh_interval)
     return {"solar_edge_api_key": solar_edge_api_key,
+            "solar_edge_site_id": solar_edge_site_id,
             "wattage_threshold": wattage_threshold,
             "refresh_interval": refresh_interval}
 
@@ -37,29 +40,31 @@ def get_device_configuration_from_file():
     logging.info("Starting to parse data from %s..." % DEVICE_DATA_FILE_NAME)
     with open(DEVICE_DATA_FILE_NAME, "r") as file:
         for line in file:
-            line_contents = line.split(",")
-            if line_contents[0] == "lp":  # To omit headers
+            row_contents = line.split(",")
+            if row_contents[0] == "no":  # To omit headers
                 continue
-            line_number = line_contents[0]
-            device_name = line_contents[1]
-            device_id = line_contents[2]
-            device_ip_address = line_contents[3]
-            device_local_key = line_contents[4]
+            row_number = row_contents[0]
+            device_name = row_contents[1]
+            device_id = row_contents[2]
+            device_ip_address = row_contents[3]
+            device_local_key = row_contents[4]
             if (device_name != "") & (device_id != "") & (device_ip_address != "") & (device_local_key != ""):
                 if len(device_ip_address.split(".")) != 4:
-                    logging.info("IP address for %s read from row %s in %s is invalid." % (device_name, line_number,
+                    logging.info("IP address for %s read from row %s in %s is invalid." % (device_name, row_number,
                                                                                            DEVICE_DATA_FILE_NAME))
                 if len(device_id) != 20:
-                    logging.info("Device ID for %s read from row %s in %s is invalid." % (device_name, line_number,
+                    logging.info("Device ID for %s read from row %s in %s is invalid." % (device_name, row_number,
                                                                                           DEVICE_DATA_FILE_NAME))
                 if len(device_local_key) != 17:
-                    logging.info("Device local key for %s read from row %s in %s is invalid." % (device_name, line_number,
-                                                                                                 DEVICE_DATA_FILE_NAME))
+                    logging.info(
+                        "Device local key for %s read from row %s in %s is invalid." % (device_name, row_number,
+                                                                                        DEVICE_DATA_FILE_NAME))
             else:
-                logging.warning("Data in row %s in %s is not complete. Skipping..." % (line_number, DEVICE_DATA_FILE_NAME))
+                logging.warning(
+                    "Data in row %s in %s is not complete. Skipping..." % (row_number, DEVICE_DATA_FILE_NAME))
                 continue
 
-            DEVICES_CONFIGURATION.append({"line_number": line_number, "device_name": device_name, "device_id": device_id,
+            DEVICES_CONFIGURATION.append({"row_number": row_number, "device_name": device_name, "device_id": device_id,
                                           "device_ip_address": device_ip_address, "device_local_key": device_local_key})
 
     logging.info("Parsing data from %s finished" % DEVICE_DATA_FILE_NAME)
@@ -70,27 +75,28 @@ def get_input_parameters():
     i = 0
     for arg in args:
         if len(args) == 1:
-            print("No argument was specified. Accepted parameters: -start, -stop, -help")
+            print("No argument was specified. Accepted parameters: --start, --stop, --help")
             sys.exit(0)
 
-        i = i + 1       # To omit MAIN.PY argument
+        i = i + 1  # To omit MAIN.PY argument
         if i == 1:
             continue
 
         arg = arg.upper()
 
-        if arg in ("START", "-START"):
+        if arg in ("START", "--START"):
             start()
-        elif arg in ("STOP", "-STOP"):
+        elif arg in ("STOP", "--STOP"):
             stop()
-        elif arg in ("HELP", "-HELP"):
-            print("Ten program to SmartBedkow. Przyjmuje parametry -start, -stop i -help.")
-            print("     Uruchomienie programu z parametrem -start spowoduje uruchomienie programu.")
-            print("     Uruchomienie programu z parametrem -stop spowoduje jego zatrzymanie.")
-            print("     Uruchomienie programu z parametrem -help spowoduje wyświetlenie komunikatów, które właśnie czytasz.")
-            print("                                                                                    - Rafał, 05/05/2022.")
+        elif arg in ("HELP", "--HELP"):
+            print("This is SmartBedkow. Accepts following parameters: --start, --stop and --help.")
+            print("     Starting application with --start parameter will start the program.")
+            print("     Starting application with --stop parameter will stop the program.")
+            print("     Starting application with --help parameter will display the message you are reading right now.")
+            print(
+                "                                                                                    - Rafał, 05/05/2022.")
         else:
-            print("Unknown argument: %s. Accepted parameters: -start, -stop, -help" % args[1])
+            print("Unknown argument: %s. Accepted parameters: --start, --stop, --help" % args[1])
             sys.exit(0)
 
 
@@ -104,10 +110,11 @@ def start():
     refresh_interval_in_minutes = refresh_interval_in_seconds / 60
     wattage_threshold = int(app_config["wattage_threshold"])
     solar_edge_api_key = app_config["solar_edge_api_key"]
+    solar_edge_site_id = app_config["solar_edge_site_id"]
 
     get_device_configuration_from_file()
 
-    logging.info("%s device(s) will be initialized" % len(DEVICES_CONFIGURATION))
+    logging.info("%s device(s) will be attempted to initialize" % len(DEVICES_CONFIGURATION))
 
     device_initialized = False
 
@@ -128,7 +135,8 @@ def start():
     else:
         logging.info("Device(s) initialization completed")
 
-    se = solar_edge.SolarEdge(solar_edge_api_key, SOLAR_EDGE_API_START_TIME, SOLAR_EDGE_API_END_TIME)
+    se = solar_edge.SolarEdge(solar_edge_api_key, solar_edge_site_id, SOLAR_EDGE_API_START_TIME,
+                              SOLAR_EDGE_API_END_TIME)
 
     logging.warning("S M A R T B E D K O W   S T A R T E D")
 
@@ -137,6 +145,12 @@ def start():
         se.get_solar_edge_api_response()
 
         se_power_production = se.get_power_production()
+        # low_temperature_power_production_override_in_celsius - read the current temperature from the radiator (?)
+        # co niesie get_status()?
+        # detect_available_dps()   - datapoints z device
+        # json['result']['properties'][3]['value']
+        # curl --request GET "https://openapi.tuyaeu.com/v2.0/cloud/thing/40631580483fda1abcab/shadow/properties?codes=Temp_current" --header "sign_method: HMAC-SHA256" --header "client_id: 4je3h7mpbj1bo1uj86zj" --header "t: 1699561886811" --header "mode: cors" --header "Content-Type: application/json" --header "sign: E89B880C97FD6F825A7D1A66F6C1B5F43E1AA80570B1D334F7BC779F18EC160F" --header "access_token: 1c8d9c658e80d0e8106bac4d3bdbe695"
+        # to sie tak da przez cloud api, ale lepiej to robic w local
         if se_power_production >= wattage_threshold:  # Turn on radiators
             for device in DEVICES:
                 if not device.is_on():
@@ -152,22 +166,34 @@ def start():
 
 
 def stop():
-    #Linux only
-    logging.info("Program stopped by " + os.getlogin())
-    os.system('ps axf | grep main | grep -v grep | awk \'{print "kill -9 " $1 }\' | sh')
+    # Linux only
+    if platform.system() == "Linux":
+        logging.info("Program stopped by " + os.getlogin())
+        os.system('ps axf | grep main | grep -v grep | awk \'{print "kill -9 " $1 }\' | sh')
+    else:
+        print("This command will not work on OS other than Linux")
 
 
 def print_smart_bedkow():
     print("")
-    print("   .-'''-. ,---.    ,---.   ____    .-------. ,---------.  _______       .-''-.   ______     .--.   .--.      ,-----.    .--.      .--. ")
-    print("  / _     \|    \  /    | .'  __ `. |  _ _   \\\          \\\  ____  \   .'_ _   \ |    _ `''. |  | _/  /     .'  .-,  '.  |  |_     |  | ")
-    print(" (`' )/`--'|  ,  \/  ,  |/   '  \  \| ( ' )  | `--.  ,---'| |    \ |  / ( ` )   '| _ | ) _  \| (`' ) /     / ,-.|  \ _ \ | _( )_   |  | ")
-    print("(_ o _).   |  |\_   /|  ||___|  /  ||(_ o _) /    |   \   | |____/ / . (_ o _)  ||( ''_'  ) ||(_ ()_)     ;  \  '_ /  | :|(_ o _)  |  | ")
-    print(" (_,_). '. |  _( )_/ |  |   _.-`   || (_,_).' __  :_ _:   |   _ _ '. |  (_,_)___|| . (_) `. || (_,_)   __ |  _`,/ \ _/  || (_,_) \ |  | ")
-    print(".---.  \  :| (_ o _) |  |.'   _    ||  |\ \  |  | (_I_)   |  ( ' )  \\'  \   .---.|(_    ._) '|  |\ \  |  |: (  '\_/ \   ;|  |/    \|  | ")
-    print("\    `-'  ||  (_,_)  |  ||  _( )_  ||  | \ `'   /(_(=)_)  | (_{;}_) | \  `-'    /|  (_.\.' / |  | \ `'   / \ `"'"/  \  ) / |  '  "'"' '" /\  `  | ")
-    print(" \       / |  |      |  |\ (_ o _) /|  |  \    /  (_I_)   |  (_,_)  /  \       / |       .'  |  |  \    /   '. \_/``"'"''.'"'""  |    /  \    | ")
-    print("  `-...-'  '--'      '--' '.(_,_).' ''-'   `'-'   '---'   /_______.'    `'-..-'  '-----'`    `--'   `'-'      '-----'    `---'    `---` ")
+    print(
+        "   .-'''-. ,---.    ,---.   ____    .-------. ,---------.  _______       .-''-.   ______     .--.   .--.      ,-----.    .--.      .--. ")
+    print(
+        "  / _     \|    \  /    | .'  __ `. |  _ _   \\\          \\\  ____  \   .'_ _   \ |    _ `''. |  | _/  /     .'  .-,  '.  |  |_     |  | ")
+    print(
+        " (`' )/`--'|  ,  \/  ,  |/   '  \  \| ( ' )  | `--.  ,---'| |    \ |  / ( ` )   '| _ | ) _  \| (`' ) /     / ,-.|  \ _ \ | _( )_   |  | ")
+    print(
+        "(_ o _).   |  |\_   /|  ||___|  /  ||(_ o _) /    |   \   | |____/ / . (_ o _)  ||( ''_'  ) ||(_ ()_)     ;  \  '_ /  | :|(_ o _)  |  | ")
+    print(
+        " (_,_). '. |  _( )_/ |  |   _.-`   || (_,_).' __  :_ _:   |   _ _ '. |  (_,_)___|| . (_) `. || (_,_)   __ |  _`,/ \ _/  || (_,_) \ |  | ")
+    print(
+        ".---.  \  :| (_ o _) |  |.'   _    ||  |\ \  |  | (_I_)   |  ( ' )  \\'  \   .---.|(_    ._) '|  |\ \  |  |: (  '\_/ \   ;|  |/    \|  | ")
+    print(
+        "\    `-'  ||  (_,_)  |  ||  _( )_  ||  | \ `'   /(_(=)_)  | (_{;}_) | \  `-'    /|  (_.\.' / |  | \ `'   / \ `"'"/  \  ) / |  '  "'"' '" /\  `  | ")
+    print(
+        " \       / |  |      |  |\ (_ o _) /|  |  \    /  (_I_)   |  (_,_)  /  \       / |       .'  |  |  \    /   '. \_/``"'"''.'"'""  |    /  \    | ")
+    print(
+        "  `-...-'  '--'      '--' '.(_,_).' ''-'   `'-'   '---'   /_______.'    `'-..-'  '-----'`    `--'   `'-'      '-----'    `---'    `---` ")
     print("")
